Remove dead code and reformat.
