--- scalafix-core/src/main/scala/scalafix/Scalafix.scala
@@ -1,16 +1,0 @@
-package scalafix
-
-import scala.meta._
-import scala.util.control.NonFatal
-
-object Scalafix {
-//  def fix(input: String, config: ScalafixConfig): Fixed =
-//    fix(Input.String(input), config)
-//  def fix(input: Input, config: ScalafixConfig): Fixed =
-//    config.dialect(input).parse[Source] match {
-//      case parsers.Parsed.Success(ast) =>
-//        fix(RewriteCtx(ast, config))
-//      case parsers.Parsed.Error(pos, msg, details) =>
-//        Fixed.Failed(Failure.ParseError(pos, msg, details))
-//    }
-}
